Updated to work correctly with the statsd module.
Do extra looking at objects so that when called from a subclass,
the name of the subclass will be sent as the stat, instead
of the superclass where the method was decorated. This is in
timer_for_subclasses and counter_for_subclasses.

If you *want* the superclass name to appear, timer and counter are
available and don't do this extra work.